requirements.txt
asgiref==3.5.2
astroid==2.11.5
backports.zoneinfo==0.2.1
dill==0.3.5.1
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.1.1
platformdirs==2.5.2
PyJWT==2.4.0
pylint==2.13.9
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
uritemplate==4.1.1
wrapt==1.14.1
